ai-backend/requirements.txt
# FastAPI core dependencies
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.2
python-multipart==0.0.9

# MongoDB (async)
motor==3.5.1
pymongo==4.5.0

# AI/ML logic
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# NLP tools
spacy==3.7.2
rapidfuzz==3.1.1
pyspellchecker==0.7.0
sentence-transformers==2.2.2
transformers==4.34.0
torch==2.2.0

# Environment management
python-dotenv==1.0.0